Update get.sources.sh
build-lool.sh is creating the directories core and online in line 41 and 42 atm. so the checks if the directories exists will always be true and the git clone won't be execeuted. I changed it to "check if core/.git" directory is available and if not run the git clone, otherwise change into and fetch.